More SleepGraphs improvements
A bug in SleepGraphs incorrectly calculated the number of flies in each
genotype, which ended up filling in a bunch of zeros and generally
making everything undesirable. There was also a problem with nans
getting included in the mean calculations, which is now fixed.

Also added new feature to graphs: standard deviations!