Misc. minor updates and fixes
* Made a few minor corrections to fix typos and delete unnecessary
comments/code

* Changed the default sizing for sleep graphs; automatic graphs should
now be easier to view without needing to change the boundaries

* Updated sample input file to be more illustrative of the current data
input system

* Commented out periodicity calculations in actogram2master. These lines
of code were causing the program to break; a future release will solve
this problem. I flagged the lines of code that involve periodicity
calculations, and left the old (commented out) code directly under the
code that excludes the periodicity analysis.